Move passwordHandling under cmd/server, update deps
Rename passwordHandling/Argon2stuff.go to cmd/server/passwordHandling/Argon2stuff.go and update the import in cmd/server/commands.go to the new path. Add github.com/hraban/opus to go.mod (deduplicate previous indirect entry). Remove the releases/v0.1.3.md release notes file.
