--- releases/v0.1.3.md
@@ -1,19 +1,0 @@
-## v0.1.3 Release Notes
-
-**This release adds official Mac and Linux builds.**
-
-- The Windows build is identical to v0.1.2 (Voice UI Fixes).
-- Now includes:
-    - `DarkWave-windows.zip` (for most Windows users)
-    - `DarkWave-mac` (Mac executable)
-    - `DarkWave-linux` (Linux executable)
-- Windows users on older PCs or virtual machines **without proper GPU drivers**: You may see a black screen. In this case, download the `DarkWave-windows-mesa-fallback.zip` (to be provided in future releases), and extract the DLLs next to `DarkWave.exe`.
-- "Source code (zip/tar.gz)" assets contain the entire source, not the client build.
-
-**Full voice UI changelog:**
-- Fixed voice members not displaying in voice panel
-- Improved voice panel UI with speaking indicators (● for speaking, ○ for silent)
-- Fixed room list sidebar to show voice channel members instantly
-- Volume control for individual peers via right-click
-
-*(v0.1.2 was a manual release for Windows only; v0.1.3 is the first release with automated cross-platform builds!)*
